backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
motor==3.3.1
python-dotenv==1.2.1
pymongo==4.5.0
pydantic==2.12.3
httpx==0.27.0